I-#430: refactored nosql models
