+++ core/services/aws/commons/ddbModels/AcceptDonationModel.ts
@@ -1,4 +1,4 @@
-import type { AcceptedDonationDTO } from '../../../../commons/dto/DonationDTO'
+import type { AcceptDonationDTO } from '../../../../../commons/dto/DonationDTO'
 import type {
   DbModelDtoAdapter,
   HasTimeLog,
@@ -11,8 +11,8 @@ import type {
 export const ACCEPTED_DONATION_PK_PREFIX = 'BLOOD_REQ'
 export const ACCEPTED_DONATION_SK_PREFIX = 'ACCEPTED'
 
-export type AcceptedDonationFields = Omit<
-AcceptedDonationDTO,
+export type AcceptDonationFields = Omit<
+AcceptDonationDTO,
 'id' | 'seekerId' | 'requestPostId' | 'donorId'
 > &
 HasTimeLog & {
@@ -22,21 +22,21 @@ HasTimeLog & {
 
 export class AcceptDonationRequestModel
 implements
-    NosqlModel<AcceptedDonationFields>,
-    DbModelDtoAdapter<AcceptedDonationDTO, AcceptedDonationFields> {
-  getIndexDefinitions(): IndexDefinitions<AcceptedDonationFields> {
+    NosqlModel<AcceptDonationFields>,
+    DbModelDtoAdapter<AcceptDonationDTO, AcceptDonationFields> {
+  getIndexDefinitions(): IndexDefinitions<AcceptDonationFields> {
     return {}
   }
 
-  getPrimaryIndex(): DbIndex<AcceptedDonationFields> {
+  getPrimaryIndex(): DbIndex<AcceptDonationFields> {
     return { partitionKey: 'PK', sortKey: 'SK' }
   }
 
-  getIndex(indexType: IndexType, indexName: string): DbIndex<AcceptedDonationFields> | undefined {
+  getIndex(indexType: IndexType, indexName: string): DbIndex<AcceptDonationFields> | undefined {
     return this.getIndexDefinitions()[indexType]?.[indexName]
   }
 
-  fromDto(acceptedDonationDto: AcceptedDonationDTO): AcceptedDonationFields {
+  fromDto(acceptedDonationDto: AcceptDonationDTO): AcceptDonationFields {
     const { seekerId, requestPostId, donorId, ...remainingData } = acceptedDonationDto
     return {
       PK: `${ACCEPTED_DONATION_PK_PREFIX}#${seekerId}`,
@@ -45,7 +45,7 @@ implements
     }
   }
 
-  toDto(dbFields: AcceptedDonationFields): AcceptedDonationDTO {
+  toDto(dbFields: AcceptDonationFields): AcceptDonationDTO {
     const { PK, SK, ...remainingFields } = dbFields
     return {
       ...remainingFields,
