+++ core/services/aws/commons/ddbModels/DonationRecordModel.ts
@@ -1,4 +1,4 @@
-import type { DonationRecordDTO } from '../../../../commons/dto/DonationDTO'
+import type { DonationRecordDTO } from '../../../../../commons/dto/DonationDTO'
 import type {
   DbModelDtoAdapter,
   HasTimeLog,
