+++ core/services/aws/commons/ddbModels/DonorSearchModel.ts
@@ -1,5 +1,5 @@
-import type { DonorSearchDTO} from '../../../../commons/dto/DonationDTO';
-import { DonationStatus } from '../../../../commons/dto/DonationDTO'
+import type { DonorSearchDTO} from '../../../../../commons/dto/DonationDTO';
+import { DonationStatus } from '../../../../../commons/dto/DonationDTO'
 import type {
   DbModelDtoAdapter,
   HasTimeLog,
